SYNRGY-134 - Service and controller layers
- Removed classes folder
- split user.js into controller and service layers
- removed the controller logic inside of the routes/user.js file
